interpreter: change truthiness mechanics once again
We used to interpret null and false as falsy, true as truthy and
anything else would cause a TypeError, however, this patch adds two
different truthiness mechanics, where logical operators (logical and,
logical or and logical not) use the less restrictive mechanic and
anything else uses the more restrictive mechanic.

The restrictive mechanic interprets false and null as falsy, true as
truthy and anything else causes a TypeError. The less restrictive
mechanic interprets false and null as falsy, anything else as truthy.

Fluet will not have a nullish coalescing operator because there is no
need for one. Logical or does not interpret 0, '', et cetera as falsy,
unlike some other languages.