everywhere: Introduce several refactors
This patch introduces several refactors to help with implementing return
statements. Includes changes to function declaration parsing and
replaces `evaluate_with_env` with the `with_env` helper, which is more
general. `with_env` takes a `FnOnce` and an `Env`, and switches to the
new environment only for the lifetime of the `FnOnce`.